backend/requirements.txt
asgiref==3.8.1
astroid==3.3.4
dill==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
PyJWT==2.9.0
pylint==3.3.1
python-dotenv==1.0.1
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
